+++ src/utils/myRecruitment/category/categoryMapping.ts
