+++ src/utils/myRecruitment/category/index.ts
@@ -1,0 +1,1 @@
+export * from './categoryMapping';
