+++ src/utils/myRecruitment/index.ts
@@ -1,2 +1,5 @@
-export * from './transformFormToRequest';
-export * from './categoryMapping';
+// Form 관련 유틸 (검증, 변환)
+export * from './form';
+
+// 카테고리 매핑 유틸
+export * from './category';
